+++ library/kotlin/src/io/envoyproxy/envoymobile/Client.kt
@@ -1,32 +1,12 @@
 package io.envoyproxy.envoymobile
 
-import java.nio.ByteBuffer
-
 interface Client {
+  /**
+   * For starting a stream
+   *
+   * @param request the request for opening a stream
+   * @param streamCallback the callback for receiving stream events
+   * @return the emitter for streaming data outward
+   */
   fun startStream(request: Request, streamCallback: StreamCallback): StreamEmitter
 }
-
-interface StreamCallback {
-  fun onHeaders(statusCode: Int, headers: Map<String, List<String>>)
-  fun onData(byteBuffer: ByteBuffer, endStream: Boolean)
-  fun onMetadata(metadata: Map<String, List<String>>, endStream: Boolean)
-  fun onTrailers(trailers: Map<String, List<String>>)
-  fun onError(error: Error)
-  fun onCanceled()
-  fun onCompletion()
-}
-
-interface StreamEmitter {
-  fun isActive(): Boolean
-  fun sendData(byteBuffer: ByteBuffer, endStream: Boolean): StreamEmitter
-  fun sendMetadata(metadata: Map<String, List<String>>, endStream: Boolean): StreamEmitter
-  
-  fun close(trailers: Map<String, List<String>>)
-  fun cancel()
-}
-
-class Error(
-    val code: Int,
-    val message: String,
-    val cause: Throwable? = null
-)
